+++ Toggl.PrimeRadiant.Realm/Models/RealmTask.cs
@@ -8,13 +8,13 @@ namespace Toggl.PrimeRadiant.Realm
     {
         public string Name { get; set; }
 
-        public int EstimatedSeconds { get; set; }
+        public long EstimatedSeconds { get; set; }
 
         public bool Active { get; set; }
 
         public DateTimeOffset At { get; set; }
 
-        public int TrackedSeconds { get; set; }
+        public long TrackedSeconds { get; set; }
 
         public RealmWorkspace RealmWorkspace { get; set; }
